examples/knowledge_graph_qa_system/requirements.txt
indexify
neo4j
google-generativeai